+++ bash/decoders/dumpvdl2.sh
@@ -7,47 +7,24 @@ source $RECEIVER_BASH_DIRECTORY/functions.sh
 
 clear
 log_project_title
-log_title_heading "Setting up the ACARSDEC decoder"
+log_title_heading "Setting up the dumpvdl2 decoder"
 log_title_message "------------------------------------------------------------------------------"
 if ! whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-              --title "ACARSDEC decoder Setup" \
-              --yesno "ACARSDEC is a multi-channels acars decoder with built-in rtl_sdr, airspy front end or sdrplay device. Since 3.0, It comes with a database backend : acarsserv to store received acars messages.\n\nWould you like to begin the setup process now?" \
+              --title "dumpvdl2 decoder Setup" \
+              --yesno "dumpvdl2 is a VDL Mode 2 message decoder and protocol analyzer.\n\nWould you like to begin the setup process now?" \
               11 78; then
     echo ""
     log_alert_heading "INSTALLATION HALTED"
     log_alert_message "Setup has been halted at the request of the user"
     echo ""
     log_title_message "------------------------------------------------------------------------------"
-    log_title_heading "ACARSDEC decoder setup halted"
+    log_title_heading "dumpvdl2 decoder setup halted"
     echo ""
     exit 1
 fi
 
 
 ## GATHER REQUIRED INFORMATION FROM THE USER
-
-log_heading "Determine the device type to build ACARSDEC for"
-
-log_message "Asking which type of device will be used by ACARSDEC"
-device=$(whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-                  --title "Device Type" \
-                  --menu "Please choose the RTL-SDR device type which is to be used by ACARSDEC." \
-                  11 78 3 \
-                  "RTL-SDR" "" \
-                  "AirSpy" "" \
-                  "SDRPlay" "" \
-                  3>&1 1>&2 2>&3)
-exit_status=$?
-if [[ $exit_status != 0 ]]; then
-    echo ""
-    log_alert_heading "INSTALLATION HALTED"
-    log_alert_message "Setup has been halted at the request of the user"
-    echo ""
-    log_title_message "------------------------------------------------------------------------------"
-    log_title_heading "ACARSDEC decoder setup halted"
-    echo ""
-    exit 1
-fi
 
 log_heading "Gather information required to configure the decoder(s)"
 
@@ -79,12 +56,38 @@ if [[ -f /usr/local/bin/dumpvdl2 ]]; then
     vdl_decoder_installed="true"
 fi
 
-if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "true" ]]; then
+
+if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "true" || "${vdl_decoder_installed}" == "true" ]]; then
     log_message "Informing the user that existing decoder(s) appears to be installed"
-    whiptail --backtitle "ACARSDEC Decoder Configuration" \
+    whiptail --backtitle "Dumpvdl2 Decoder Configuration" \
              --title "RTL-SDR Dongle Assignments" \
              --msgbox "It appears that existing decoder(s) have been installed on this device. In order to run ACARSDEC in tandem with other decoders you will need to specifiy which RTL-SDR dongle each decoder is to use.\n\nKeep in mind in order to run multiple decoders on a single device you will need to have multiple RTL-SDR devices connected to your device." \
              12 78
+
+    if [[ "${acars_decoder_installed}" == "true" ]]; then
+        log_message "Determining which device is currently assigned to the UAT decoder"
+        exec_start=`get_config "ExecStart" "/etc/systemd/system/acarsdec.service"`
+        current_acars_device_number=`echo $exec_start | grep -o -P '(?<=-r ).*(?= -A)'`
+        log_message "Asking the user to assign a RTL-SDR device number to ACARSDEC"
+        acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number"
+        while [[ -z $acars_device_number ]] ; do
+            acars_device_number=$(whiptail --backtitle "ACARSDEC Decoder Configuration" \
+                                           --title "${acars_device_number_title}" \
+                                           --inputbox "\nEnter the RTL-SDR device number to assign your ACARSDEC decoder." \
+                                           8 78 \
+                                           "${current_acars_device_number}" 3>&1 1>&2 2>&3)
+            exit_status=$?
+            if [[ $exit_status != 0 ]]; then
+                log_alert_heading "INSTALLATION HALTED"
+                log_alert_message "Setup has been halted due to lack of required information"
+                echo ""
+                log_title_message "------------------------------------------------------------------------------"
+                log_title_heading "Dumpvdl2 decoder setup halted"
+                exit 1
+            fi
+            acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number (REQUIRED)"
+        done
+    fi
 
     if [[ "${adsb_decoder_installed}" == "true" ]]; then
         log_message "Determining which device is currently assigned to the ADS-B decoder"
@@ -103,7 +106,7 @@ if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "tr
                 log_alert_message "Setup has been halted due to lack of required information"
                 echo ""
                 log_title_message "------------------------------------------------------------------------------"
-                log_title_heading "ACARSDEC decoder setup halted"
+                log_title_heading "Dumpvdl2 decoder setup halted"
                 exit 1
             fi
             adsb_device_number_title="Enter the ADS-B Decoder RTL-SDR Device Number (REQUIRED)"
@@ -128,93 +131,68 @@ if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "tr
                 log_alert_message "Setup has been halted due to lack of required information"
                 echo ""
                 log_title_message "------------------------------------------------------------------------------"
-                log_title_heading "ACARSDEC decoder setup halted"
+                log_title_heading "Dumpvdl2 decoder setup halted"
                 exit 1
             fi
             uat_device_number_title="Enter the UAT Decoder RTL-SDR Device Number (REQUIRED)"
         done
     fi
 
-    if [[ "${vdl_decoder_installed}" == "true" ]]; then
+    current_vdl_device_number=""
+    if [[ "${acars_decoder_installed}" == "true" ]]; then
         log_message "Determining which device is currently assigned to the VDL decoder"
         exec_start=`get_config "ExecStart" "/etc/systemd/system/dumpvdl2.service"`
         current_vdl_device_number=`echo $exec_start | grep -o -P '(?<=--rtlsdr ).*(?= --gain)'`
-        log_message "Asking the user to assign a RTL-SDR device number to the VDL decoder"
-        vdl_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number"
-        while [[ -z $vdl_device_number ]]; do
-            vdl_device_number=$(whiptail --backtitle "Dumpvdl2 Decoder Configuration" \
-                                         --title "${vdl_device_number_title}" \
-                                         --inputbox "Enter the RTL-SDR device number to assign your dumpvdl2 decoder." \
-                                         8 78 \
-                                         "${current_vdl_device_number}" 3>&1 1>&2 2>&3)
-            exit_status=$?
-            if [[ $exit_status != 0 ]]; then
-                log_alert_heading "INSTALLATION HALTED"
-                log_alert_message "Setup has been halted due to lack of required information"
-                echo ""
-                log_title_message "------------------------------------------------------------------------------"
-                log_title_heading "Dumpvdl2 decoder setup halted"
-                exit 1
-            fi
-            vdl_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number (REQUIRED)"
-        done
     fi
-
-    current_acars_device_number=""
-    if [[ "${acars_decoder_installed}" == "true" ]]; then
-        log_message "Determining which device is currently assigned to the ACARS decoder"
-        exec_start=`get_config "ExecStart" "/etc/systemd/system/acarsdec.service"`
-        current_acars_device_number=`echo $exec_start | grep -o -P '(?<=-r ).*(?= -A)'`
-    fi
-    log_message "Asking the user to assign a RTL-SDR device number to ACARSDEC"
-    acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number"
-    while [[ -z $acars_device_number ]]; do
-        acars_device_number=$(whiptail --backtitle "ACARSDEC Decoder Configuration" \
-                                       --title "${acars_device_number_title}" \
-                                       --inputbox "Enter the RTL-SDR device number to assign your ACARSDEC decoder." \
-                                       8 78 \
-                                       "${current_acars_device_number}" 3>&1 1>&2 2>&3)
+    log_message "Asking the user to assign a RTL-SDR device number to the VDL decoder"
+    vdl_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number"
+    while [[ -z $vdl_device_number ]]; do
+        vdl_device_number=$(whiptail --backtitle "Dumpvdl2 Decoder Configuration" \
+                                     --title "${vdl_device_number_title}" \
+                                     --inputbox "Enter the RTL-SDR device number to assign your dumpvdl2 decoder." \
+                                     8 78 \
+                                     "${current_vdl_device_number}" 3>&1 1>&2 2>&3)
         exit_status=$?
         if [[ $exit_status != 0 ]]; then
             log_alert_heading "INSTALLATION HALTED"
             log_alert_message "Setup has been halted due to lack of required information"
             echo ""
             log_title_message "------------------------------------------------------------------------------"
-            log_title_heading "ACARSDEC decoder setup halted"
+            log_title_heading "Dumpvdl2 decoder setup halted"
             exit 1
         fi
-        acars_device_number_title="Enter the ACARSDEC RTL-SDR Device Number (REQUIRED)"
+        vdl_device_number_title="Enter the dumpvdl2 RTL-SDR Device Number (REQUIRED)"
     done
 fi
 
-if [[ -z $acars_device_number ]]; then
-    acars_device_number="0"
+if [[ -z $vdl_device_number ]]; then
+    vdl_device_number="0"
 fi
 
-current_acars_frequencies="130.025 130.425 130.450 131.125 131.550"
-if [[ "${acars_decoder_installed}" == "true" ]]; then
+current_vdl_frequencies="136725000 136975000 136875000"
+if [[ "${vdl_decoder_installed}" == "true" ]]; then
     log_message "Determining which frequencies are currently assigned"
-    exec_start=`get_config "ExecStart" "/etc/systemd/system/acarsdec.service"`
-    current_acars_frequencies=`sed -e "s#.*-r ${acars_device_number} \(\)#\1#" <<< "${exec_start}"`
+    exec_start=`get_config "ExecStart" "/etc/systemd/system/dumpvdl2.service"`
+    current_vdl_frequencies=`sed -e "s#.*--correction ${vdl_correction} \(\)#\1#" <<< "${exec_start}"`
 fi
-log_message "Asking the user for ACARS frequencies to monitor"
-acars_fequencies_title="Enter ACARS Frequencies"
+log_message "Asking the user for VDL frequencies to monitor"
+acars_fequencies_title="Enter VDL Frequencies"
 while [[ -z $acars_fequencies ]] ; do
-    acars_fequencies=$(whiptail --backtitle "ACARS Frequencies" \
-                                --title "${acars_fequencies_title}" \
-                                --inputbox "\nEnter the ACARS frequencies you would like to monitor." \
-                                8 78 \
-                                "${current_acars_frequencies}" 3>&1 1>&2 2>&3)
+    vdl_fequencies=$(whiptail --backtitle "VDL Frequencies" \
+                              --title "${vdl_fequencies_title}" \
+                              --inputbox "\nEnter the VDL frequencies you would like to monitor." \
+                              8 78 \
+                              "${current_vdl_frequencies}" 3>&1 1>&2 2>&3)
     exit_status=$?
     if [[ $exit_status != 0 ]]; then
         log_alert_heading "INSTALLATION HALTED"
         log_alert_message "Setup has been halted due to lack of required information"
         echo ""
         log_title_message "------------------------------------------------------------------------------"
-        log_title_heading "ACARSDEC decoder setup halted"
+        log_title_heading "Dumpvdl2 decoder setup halted"
         exit 1
     fi
-    acars_fequencies_title="Enter ACARS Frequencies (REQUIRED)"
+    vdl_fequencies_title="Enter VDL Frequencies (REQUIRED)"
 done
 
 
@@ -222,13 +200,16 @@ done
 
 log_heading "Installing packages needed to fulfill dependencies for FlightAware PiAware client"
 
+check_package build-essential
 check_package cmake
+check_package git
+check_package libglib2.0-dev
 check_package libjansson-dev
-check_package libpaho-mqtt-dev
-check_package libsndfile1-dev
+check_package libprotobuf-c-dev
+check_package librtlsdr-dev
 check_package libsqlite3-dev
-check_package libusb-1.0-0-dev
 check_package libxml2-dev
+check_package libzmq3-dev
 check_package pkg-config
 check_package zlib1g-dev
 
@@ -243,18 +224,6 @@ case $RECEIVER_OS_DISTRIBUTION in
 esac
 check_package sqlite3
 check_package php${distro_php_version}-sqlite3
-
-case "${device}" in
-    "RTL-SDR")
-        check_package librtlsdr-dev
-        ;;
-    "AirSpy")
-        check_package libairspy-dev
-        ;;
-    "SDRPlay")
-        check_package libmirisdr-dev
-        ;;
-esac
 
 
 ## BLACKLIST UNWANTED RTL-SDR MODULES
@@ -315,57 +284,46 @@ log_message "Running ldconfig"
 sudo ldconfig
 
 
-## CLONE OR PULL THE ACARSDEC GIT REPOSITORY
+## CLONE OR PULL THE DUMPVDL2 GIT REPOSITORY
 
-log_heading "Preparing the ACARSDEC Git repository"
+log_heading "Preparing the dumpvdl2 Git repository"
 
-if [[ -d $RECEIVER_BUILD_DIRECTORY/acarsdec && -d $RECEIVER_BUILD_DIRECTORY/acarsdec/.git ]]; then
-    log_message "Entering the ACARSDEC git repository directory"
-    cd $RECEIVER_BUILD_DIRECTORY/acarsdec
-    log_message "Updating the local ACARSDEC git repository"
+if [[ -d $RECEIVER_BUILD_DIRECTORY/dumpvdl2 && -d $RECEIVER_BUILD_DIRECTORY/dumpvdl2/.git ]]; then
+    log_message "Entering the dumpvdl2 git repository directory"
+    cd $RECEIVER_BUILD_DIRECTORY/dumpvdl2
+    log_message "Updating the local dumpvdl2 git repository"
     echo ""
     git pull
 else
     log_message "Entering the build directory"
     cd $RECEIVER_BUILD_DIRECTORY
-    log_message "Cloning the ACARSDEC git repository locally"
+    log_message "Cloning the dumpvdl2 git repository locally"
     echo ""
-    git clone https://github.com/TLeconte/acarsdec.git
+    git clone https://github.com/szpajder/dumpvdl2.git
 fi
 
 
-## BUILD AND INSTALL THE ACARSDEC BINARY
+## BUILD AND INSTALL THE DUMPVDL2 BINARY
 
-log_heading "Building the ACARSDEC binary"
+log_heading "Building the dumpvdl2 binary"
 
-if [[ ! -d $RECEIVER_BUILD_DIRECTORY/acarsdec/build ]]; then
-    log_message "Creating the ACARSDEC build directory"
+if [[ ! -d $RECEIVER_BUILD_DIRECTORY/dumpvdl2/build ]]; then
+    log_message "Creating the dumpvdl2 build directory"
     echo ""
-        mkdir -vp $RECEIVER_BUILD_DIRECTORY/acarsdec/build
+        mkdir -vp $RECEIVER_BUILD_DIRECTORY/dumpvdl2/build
     echo ""
 fi
-if [[ -n "$(ls -A $RECEIVER_BUILD_DIRECTORY/acarsdec/build 2>/dev/null)" ]]; then
-    log_message "Deleting all files currently residing in the ACARSDEC build directory"
-    rm -rf $RECEIVER_BUILD_DIRECTORY/acarsdec/build/*
+if [[ -n "$(ls -A $RECEIVER_BUILD_DIRECTORY/dumpvdl2/build 2>/dev/null)" ]]; then
+    log_message "Deleting all files currently residing in the dumpvdl2 build directory"
+    rm -rf $RECEIVER_BUILD_DIRECTORY/dumpvdl2/build/*
 fi
-log_message "Entering the ACARSDEC build directory"
-cd $RECEIVER_BUILD_DIRECTORY/acarsdec/build
+log_message "Entering the dumpvdl2 build directory"
+cd $RECEIVER_BUILD_DIRECTORY/dumpvdl2/build
 
 log_message "Executing cmake"
 echo ""
-case "${device}" in
-    "RTL-SDR")
-        cmake .. -Drtl=ON
-        ;;
-    "AirSpy")
-        cmake .. -Dairspy=ON
-        ;;
-    "SDRPlay")
-        cmake .. -Dsdrplay=ON
-        ;;
-esac
+cmake ../
 echo ""
-
 log_message "Executing make"
 echo ""
 make
@@ -375,47 +333,16 @@ echo ""
 sudo make install
 
 
-## CLONE OR PULL THE ACARSSERV GIT REPOSITORY
+## RUN DUMPVDL2
 
-log_heading "Preparing the acarsserv Git repository"
-
-if [[ -d $RECEIVER_BUILD_DIRECTORY/acarsserv && -d $RECEIVER_BUILD_DIRECTORY/acarsserv/.git ]]; then
-    log_message "Entering the acarsserv git repository directory"
-    cd $RECEIVER_BUILD_DIRECTORY/acarsserv
-    log_message "Updating the local acarsserv git repository"
-    echo ""
-    git pull
-else
-    log_message "Entering the build directory"
-    cd $RECEIVER_BUILD_DIRECTORY
-    log_message "Cloning the acarsserv git repository locally"
-    echo ""
-    git clone https://github.com/TLeconte/acarsserv.git
-fi
-
-
-## BUILD AND INSTALL THE ACARSSERV BINARY
-
-log_heading "Building the ACARSSERV binary"
-
-log_message "Entering the acarsserv build directory"
-cd $RECEIVER_BUILD_DIRECTORY/acarsserv
-log_message "Executing make"
-echo ""
-make -f Makefile
-echo ""
-
-
-## RUN ACARSDECO AND ACARSSERV
-
-log_message "Creating the ACARSDEC systemd service script"
-sudo tee /etc/systemd/system/acarsdec.service > /dev/null <<EOF
+log_message "Creating the dumpvdl2 systemd service script"
+sudo tee /etc/systemd/system/dumpvdl2.service > /dev/null <<EOF
 [Unit]
-Description=ARCARSDEC multi-channel acars decoder.
+Description=Dumpvdl2 VDL Mode 2 message decoder and protocol analyzer.
 After=network.target
 
 [Service]
-ExecStart=/usr/local/bin/acarsdec -j 127.0.0.1:5555 -o2 -g280 -r 0 130.025 130.425 130.450 131.125 131.550
+ExecStart=/usr/local/bin/dumpvdl2 --rtlsdr 0 --gain 40 --correction 42 136.975
 WorkingDirectory=/usr/local/bin
 StandardOutput=null
 TimeoutSec=30
@@ -428,30 +355,10 @@ StartLimitBurst=10
 WantedBy=multi-user.target
 EOF
 
-log_message "Creating the ACARSSERV systemd service script"
-sudo tee /etc/systemd/system/acarsserv.service > /dev/null <<EOF
-[Unit]
-Description=ARCARSSERV saves acars data to SQLite.
-After=network.target
-
-[Service]
-ExecStart=${RECEIVER_BUILD_DIRECTORY}/acarsserv/acarsserv -j 127.0.0.1:5555
-WorkingDirectory=${RECEIVER_BUILD_DIRECTORY}/acarsserv
-StandardOutput=null
-TimeoutSec=30
-Restart=on-failure
-RestartSec=30
-StartLimitInterval=350
-StartLimitBurst=10
-
-[Install]
-WantedBy=multi-user.target
-EOF
-
-log_message "Enabling then starting the ACARSDEC service"
-sudo systemctl enable --now acarsdec.service
-log_message "Enabling then starting the acarsserv service"
-sudo systemctl enable --now acarsserv.service
+log_message "Enabling then starting the dumpvdl2 service"
+sudo systemctl enable --now dumpvdl2.service
+log_message "Enabling then starting the dumpvdl2 service"
+sudo systemctl enable --now dumpvdl2.service
 
 
 ## CONFIGURATION
@@ -459,6 +366,15 @@ sudo systemctl enable --now acarsserv.service
 if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "true" ]]; then
 
     log_heading "Configuring the decoders so they can work in tandem"
+
+    if [[ "${acars_decoder_installed}" == "true" ]]; then
+        log_message "Assigning RTL-SDR device number ${acars_device_number} to ACARSDEC"
+        sudo sed -i -e "s/\(.*-r \)\([0-9]\+\)\( .*\)/\1${acars_device_number}\3/g" /etc/systemd/system/acarsdec.service
+        log_message "Reload systemd units"
+        sudo systemctl daemon-reload
+        log_message "Restarting ACARSDEC"
+        sudo systemctl restart acarsdec
+    fi
 
     if [[ "${adsb_decoder_installed}" == "true" ]]; then
         log_message "Assigning RTL-SDR device number ${adsb_device_number} to the FlightAware Dump1090 decoder"
@@ -474,29 +390,20 @@ if [[ "${adsb_decoder_installed}" == "true" || "${uat_decoder_installed}" == "tr
         sudo systemctl restart dump978-fa
     fi
 
-    if [[ "${vdl_decoder_installed}" == "true" ]]; then
-        log_message "Assigning RTL-SDR device number ${vdl_device_number} to dumpvdl2"
-        sudo sed -i -e "s|\(.*--rtlsdr \)\([0-9]\+\)\( .*\)|\1${vdl_device_number}\3|g" /etc/systemd/system/dumpvdl2.service
-        log_message "Reloading systemd units"
-        sudo systemctl daemon-reload
-        log_message "Restarting dumpvdl2"
-        sudo systemctl restart dumpvdl2
-    fi
-
-    log_message "Assigning RTL-SDR device number ${acars_device_number} to ACARSDEC"
-    sudo sed -i -e "s|\(.*-r \)\([0-9]\+\)\( .*\)|\1${acars_device_number}\3|g" /etc/systemd/system/acarsdec.service
+    log_message "Assigning RTL-SDR device number ${vdl_device_number} to dumpvdl2"
+    sudo sed -i -e "s|\(.*--rtlsdr \)\([0-9]\+\)\( .*\)|\1${vdl_device_number}\3|g" /etc/systemd/system/dumpvdl2.service
     log_message "Reloading systemd units"
     sudo systemctl daemon-reload
-    log_message "Restarting ACARSDEC"
-    sudo systemctl restart acarsdec
+    log_message "Restarting dumpvdl2"
+    sudo systemctl restart dumpvdl2
 fi
 
 
 ## POST INSTALLATION OPERATIONS
 
 whiptail --backtitle "${RECEIVER_PROJECT_TITLE}" \
-         --title "ACARSDEC Decoder Setup Complete" \
-         --msgbox "The setup process currently sets basic parameters needed to feed acarsserv. You can fine tune your installation by modifying the startup command found in the file /etc/systemd/system/acarsdec.service. Usage information for ACARSDEC can be found in the projects README at https://github.com/TLeconte/acarsdec." \
+         --title "Dumpvdl2 Decoder Setup Complete" \
+         --msgbox "The setup process currently sets basic parameters needed to run a basic dumpvdl2 setup. You can fine tune your installation by modifying the startup command found in the file /etc/systemd/system/dumpvdl2.service. Usage information for dumpvdl2 can be found in the projects README at https://github.com/szpajder/dumpvdl2." \
          12 78
 
 
@@ -507,7 +414,7 @@ cd $RECEIVER_ROOT_DIRECTORY
 
 echo ""
 log_title_message "------------------------------------------------------------------------------"
-log_title_heading "ACARSDEC decoder setup is complete"
+log_title_heading "Dumpvdl2 decoder setup is complete"
 echo ""
 read -p "Press enter to continue..." discard